Rebuild the entire library from the ground up
This is my attempt at getting development back on track with good source version control.

I've gone back and forth on the structure of this library a lot, and I think I'm settled on a monolith now.
I'm not going to be writing any amazing new crc checksum, compression APIs or hash table implementations - the extraneous stuff
I've been working on is useful only in so far as it's easily available to the library itself and the users of it.
Splitting it into many libraries breaks that - and I'd rather start bringing in big dependencies instead at that point.

I hope to keep things sanitary enough that components can be split out in future. We'll see how well that goes.

Doxygen WIP